merengueproj/requirements.txt
# python libraries
PIL
BeautifulSoup==3.2.0
encutils
feedparser
# for twitter plugin
python-twitter==0.8.2

# Django
Django==1.3

# Django applications
south==0.7.3
django-extensions==0.5
cssutils
django-tagging>=0.3.1
django-pagination
johnny-cache>=0.3.0
template_utils
django-debug-toolbar
django-mptt
django-oembed
django-ajax-selects
django-threadedcomments==0.5.3
django-notification==0.1.5
django-announcements>=0.1.2
django-oot
django-genericforeignkey>=0.60.4
django-stdfile
cmsutils>=0.6.1
transhette>=0.5.10
django-inlinetrans>=0.4.10dev
django-transmeta>=0.6.1
sorl-thumbnail>=11.01
searchform>=0.1.3dev
django-configfield>=0.1.0-beta
django-form-admin>=0.3
django-autoreports>=0.8.2
django_compressor==0.5.3
django-classy-tags==0.3.0
django-inplaceedit>=0.74
